--- src/deneme_neslihan.java
@@ -1,5 +1,0 @@
-public class deneme_neslihan {
-    public static void main(String[] args) {
-
-    }
-}
